feat: resolveSigningKeyRefsFromState util function moved to handler & output adjustments
